Rename common folder to lib
