+++ packages/react-composites/tests/README.md
@@ -109,7 +109,7 @@ For a primer on conditional compilation in this repository, see the [top-level d
 
 Just like the rest of the UI library code, the test applications for browser tests use conditional compilation. By contrast, the browser tests themselves are not conditionally compiled (yet!).
 
-For now, you must explicitly skip tests that are not relevant to stable flavored builds via [`isTestProfileStableFlavor()`](./browser/common/utils.ts)
+For now, you must explicitly skip tests that are not relevant to stable flavored builds via [`isTestProfileStableFlavor()`](./browser/lib/utils.ts)
 
 ```TypeScript
 test.skip(isTestProfileStableFlavor());
@@ -147,6 +147,6 @@ To debug a particular test,
 * Single-step through the test, record a video etc through the Inspector.
 
 ### Running just specific hermetic tests
-When working on a specific hermetic test be sure to use `node scripts/runBrowserTests.mjs -l`. This should be used because all of the live tests will be run after even though you are using `test.only(...)`. This happens because the wrapper script calls playwright multiple times depending on which flags are given. 
+When working on a specific hermetic test be sure to use `node scripts/runBrowserTests.mjs -l`. This should be used because all of the live tests will be run after even though you are using `test.only(...)`. This happens because the wrapper script calls playwright multiple times depending on which flags are given.
 
 As well you will need to also use the flags `... -c call` and `... -c chat` to run just the composite tests that you are writing tests for.
