+++ packages/react-composites/tests/browser/call/app/index.tsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { initializeIconsForUITests } from '../../common/testAppUtils';
+import { initializeIconsForUITests } from '../../lib/testAppUtils';
 import { HermeticApp, shouldLoadHermeticApp } from './HermeticApp';
 import { LiveApp } from './LiveApp';
 import { parseQueryArgs } from './QueryArgs';
