+++ packages/react-composites/tests/browser/call/hermetic/RoomsCallScreen.test.ts
@@ -3,8 +3,8 @@
 
 import { buildUrlWithMockAdapter, defaultMockCallAdapterState, defaultMockRemoteParticipant, test } from './fixture';
 import { expect } from '@playwright/test';
-import { dataUiId, pageClick, stableScreenshot, waitForSelector } from '../../common/utils';
-import { IDS } from '../../common/constants';
+import { dataUiId, pageClick, stableScreenshot, waitForSelector } from '../../lib/utils';
+import { IDS } from '../../lib/constants';
 
 test.describe('Rooms DeviceButton tests for different roles', async () => {
   test('All devices are shown for Presenter', async ({ page, serverUrl }) => {
