+++ packages/react-composites/tests/browser/call/live/fixture.ts
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { TEST_PARTICIPANTS } from '../../common/constants';
-import { createCallUsers, usePagePerParticipantWithCallPermissions } from '../../common/fixtureHelpers';
-import { WorkerFixture, CallUserType } from '../../common/fixtureTypes';
+import { TEST_PARTICIPANTS } from '../../lib/constants';
+import { createCallUsers, usePagePerParticipantWithCallPermissions } from '../../lib/fixtureHelpers';
+import { WorkerFixture, CallUserType } from '../../lib/fixtureTypes';
 import { createTestServer } from '../../../server';
 import { test as base } from '@playwright/test';
 import path from 'path';
