+++ packages/react-composites/tests/browser/chat/app/LiveTestApp.tsx
@@ -14,8 +14,8 @@ import {
   useAzureCommunicationChatAdapter
 } from '../../../../src';
 // eslint-disable-next-line no-restricted-imports
-import { IDS } from '../../common/constants';
-import { verifyParamExists } from '../../common/testAppUtils';
+import { IDS } from '../../lib/constants';
+import { verifyParamExists } from '../../lib/testAppUtils';
 
 const urlSearchParams = new URLSearchParams(window.location.search);
 const params = Object.fromEntries(urlSearchParams.entries());
