+++ packages/react-composites/tests/browser/chat/app/webpack.config.js
@@ -4,7 +4,7 @@
 
 module.exports = () => {
   const babelConfig = require('../../../../.babelrc.js');
-  const commonConfig = require('../../common/testapp.webpack.config')(__dirname, babelConfig);
+  const commonConfig = require('../../lib/testapp.webpack.config')(__dirname, babelConfig);
 
   return commonConfig;
 };
