+++ packages/react-composites/tests/browser/chat/live/ChatComposite.test.ts
@@ -2,13 +2,8 @@
 // Licensed under the MIT license.
 
 import { expect } from '@playwright/test';
-import {
-  chatTestSetup,
-  sendMessage,
-  waitForMessageSeen,
-  waitForMessageWithContent
-} from '../../common/chatTestHelpers';
-import { stubMessageTimestamps, waitForChatCompositeToLoad } from '../../common/utils';
+import { chatTestSetup, sendMessage, waitForMessageSeen, waitForMessageWithContent } from '../../lib/chatTestHelpers';
+import { stubMessageTimestamps, waitForChatCompositeToLoad } from '../../lib/utils';
 import { test } from './fixture';
 
 test.describe('Chat Composite E2E Tests', () => {
