+++ packages/react-composites/tests/browser/lib/fonts/segoeui-bold.woff2
