+++ packages/react-composites/tests/browser/lib/fonts/segoeui-regular.woff2
