+++ packages/react-composites/tests/browser/lib/fonts/segoeui-semibold.woff2
