+++ packages/react-composites/tests/browser/lib/testapp.webpack.config.js
